+++ test/types.test.tsx
@@ -4,6 +4,7 @@ import withImg = require('../plugin')
 declare const image: NextImgData
 
 ;<Picture src={image} alt='Coffee' priority />
+;<Picture src={image} alt='Coffee' preload />
 ;<Picture src={image} />
 ;<Picture
   sources={[{ src: image, media: '(max-width: 767px)' }, { src: image }]}
@@ -14,6 +15,7 @@ declare const image: NextImgData
 withImg({
   nextImg: {
     formats: ['avif', 'webp'],
+    maxBareImportSize: 2048,
     cache: { mode: 'read-only', dir: 'resources', version: 'photos-v2' },
   },
 })